==> src/1. preprocessing/1.rg_csv_cleaning_and_mapping.ipynb <==
import pathlib
import pandas as pd
import numpy as np
from tqdm import tqdm
import os
import guitarpro as pygp
import glob
from aux_fun import ultimate_get_filename_2

path_to_general_read_folder = "..\..\data\DadaGP\DadaGP-v1.1"
path_to_general_write_folder = "..\..\data\BGTG\BGTG-v0.1"


df_rg = pd.read_csv('..\\..\\data\\function_guitar-ext-with-predictions.csv')
df_rg = df_rg[['Fichier', 'Track_Name', 'Partie', 'Instrument', 'Mesure', 'rg-estimation']]
display(df_rg.head())


DADAGPPATH = pathlib.Path(path_to_general_read_folder)

all_paths = list(DADAGPPATH.rglob('**/*.pygp.gp5'))
# retrieve ALL filenames in DadaGP 
all_filenames = [f.name for f in all_paths]
all_filenames_no_ext = [f.name[:-13] for f in all_paths]

index = np.random.randint(0, len(all_paths))

all_paths[index], all_filenames[index], all_filenames_no_ext[index]
series_paths = pd.Series(all_paths)
series_filenames = pd.Series(all_filenames_no_ext)

tqdm.pandas()
# df.reset_index(drop=True, inplace=True)
# drop Unnamed: 0
df_to_process = df_rg.drop_duplicates(subset=['Fichier']).copy()[['Fichier', 'Partie']]
df_to_process.loc[:, 'File_Name'] = 'Not found'
df_to_process.loc[:, 'Dadagp_Path'] = 'Not found'
df_to_process.head()

# df_to_process = df_to_process.progress_apply(ultimate_get_filename,
#                                              axis=1, args=(all_filenames_no_ext, all_paths))
# 2min with lists, 7min with Series for function 1
df_to_process = df_to_process.progress_apply(ultimate_get_filename_2,
                                             axis=1, args=(series_filenames, series_paths))
# ~5 hours if we have to check the parts... (function 2)

# Merge back
df_rg = df_rg.merge(df_to_process[['Fichier', 'File_Name', 'Dadagp_Path']], on='Fichier', how='left')

# Count number of errors
display(df_rg[df_rg['Dadagp_Path'] == 'error']['Fichier'].unique())
# Windows doesn't accept files starting with a blank space, we have to rename those

# Remove the duplicates from the dataframe
df_rg = df_rg[df_rg['Dadagp_Path'] != 'error']

display(df_rg.head())
df_rg.shape # 1358 measures removed because of the duplicates (very little)

df_rg.to_csv('../../data/df_rg_1.csv', index=False)
display(df_rg.head())
df_rg['File_Name'].nunique()
