==> src/1. preprocessing/2.mapping_dadagp_inst.ipynb <==
import pathlib
import pandas as pd
import numpy as np
from tqdm import tqdm
import os
import guitarpro as pygp
import glob

path_to_general_read_folder = "..\..\data\DadaGP\DadaGP-v1.1"
path_to_general_write_folder = "..\..\data\BGTG\BGTG-v0.1"

df_rg = pd.read_csv('../../data/df_rg_1.csv')

df_rg.head()

df_inst_part = df_rg.drop_duplicates(subset=['Dadagp_Path', 'Partie'])
# This removes the measures, conserving only one row for each part of each song
df_inst_part.shape[0], df_rg.shape[0], df_inst_part['Dadagp_Path'].nunique()
# Total number of parts, total number of rows, total number of songs

total_df_track = pd.DataFrame()

# For each track we add the DadaGP_name column
for path in tqdm(df_inst_part['Dadagp_Path'].unique()):
    df_track = df_inst_part[df_inst_part['Dadagp_Path'] == path].copy() # We copy to avoid SettingWithCopyWarning
    
    dada_gp_path = df_track['Dadagp_Path'].values[0]
    
    try:
        gp_file = pygp.parse(dada_gp_path)
        
    except:
        continue
    
    # No need for this anymore normally as the path are retrieved from DadaGP
    # try:
    #     gp_file = pygp.parse(dada_gp_path)
    # except:
    #     print('Error with file:', dada_gp_path)
    #     continue
        
    inst_list = []
    for part in gp_file.tracks:
        part_name = part.name
        
        if part_name not in df_track['Partie'].values:
            continue
        inst_name = df_track[df_track['Partie'] == part_name]['Instrument'].values[0]
        inst_list.append((part_name, inst_name))
    
    # Create the Dadagp_Name column filled with 'not_filled_yet'
    df_track['Dadagp_Name'] = 'not_filled_yet'
    nb_clean = 0
    nb_distorted = 0
    for part_inst_tuple in inst_list:
        part = part_inst_tuple[0]
        inst = part_inst_tuple[1]
        if inst == 'clean':
            df_track.loc[df_track['Partie'] == part, 'Dadagp_Name'] = 'clean' + str(nb_clean)
            nb_clean += 1
        elif inst == 'distorted':
            df_track.loc[df_track['Partie'] == part, 'Dadagp_Name'] = 'distorted' + str(nb_distorted)
            nb_distorted += 1
        else:
            df_track.loc[df_track['Partie'] == part, 'Dadagp_Name'] = inst
            
    # if 'not_filled_yet' in df_track['Dadagp_Name'].unique():
        # display(df_track)
        # print(inst_list)
    
    # Join df_track with df_inst_part
    total_df_track = pd.concat([total_df_track, df_track])

df_inst_part.shape, total_df_track.shape

# The method is correct but takes a long time to compute (1h)

path_to_test = total_df_track[total_df_track['File_Name'] == '311 - Homebrew (2)']['Dadagp_Path'].values

path_to_test


gp_file = pygp.parse(path_to_test)
    
inst_list = []
for part in gp_file.tracks:
    part_name = part.name
    print(part_name)
    
    if part_name not in df_track['Partie'].values:
        continue
    inst_name = df_track[df_track['Partie'] == part_name]['Instrument'].values[0]
    inst_list.append((part_name, inst_name))

total_df_track['Dadagp_Name'].unique()
# We should not have any 'not filled yet'
# 'not filled yet' values correspond to songs where we lost the path.
# This happens when several songs have the same name but different extensions (gp3, gp4, gp5)
# In notebook 1 we simply take the first path of the name, which is not always the right one.
# As the computation necessary to assign the correct path is extremely long, we will simply drop these songs.

# Merge the DadaGP_name back

new_df_rg = df_rg.merge(total_df_track[['Dadagp_Path', 'Partie', 'Dadagp_Name']],
                                      on=['Dadagp_Path', 'Partie'], how='left')



list_test = ['distorted0', 'clean0', 'clean1', 'bass', 'distorted1',
       'distorted2', 'distorted3', 'clean2', 'clean3', 'clean4',
       'distorted4', 'distorted5']

df_rg_2 = new_df_rg[new_df_rg['Dadagp_Name'].isin(list_test)]

df_rg_2['Dadagp_Name'].unique(), df_rg_2['File_Name'].nunique()

# Some dupicated songs have not been removed using this method because they have the same name AND the same parts' names
# in the gp file. To remove them, we perform a drop_duplicates on the Dadagp_Name, Mesure and Dadagp_Path columns.
initial_shape = df_rg_2.shape[0]
print(df_rg_2.shape, df_rg_2['Dadagp_Path'].nunique())
df_rg_2 = df_rg_2.drop_duplicates(subset=['Dadagp_Name', 'Mesure', 'Dadagp_Path'])
print(df_rg_2.shape, df_rg_2['Dadagp_Path'].nunique())

lost_rows = initial_shape - df_rg_2.shape[0]
lost_rows

df_rg_2.to_csv('../../data/df_rg_2.csv', index=False)
