==> src/1. preprocessing/4.retrieving_rythmic_and_bass_tokens.ipynb <==
# Import the autoreload extension
%load_ext autoreload

# Set autoreload to reload all modules (except those excluded by %aimport)
%autoreload 2

from aux_fun import get_tokens_inst_iter_folders, get_tokens_inst_iter_df_rg
import pandas as pd
import numpy as np
from tqdm import tqdm
import os
import guitarpro as pygp
import glob
import pathlib

path_to_general_read_folder = "..\..\data\DadaGP\DadaGP-v1.1"
path_to_general_write_folder = "..\..\data\BGTG\BGTG_RG"
windows11 = True # Some paths are different in Windows 11

### Arbitrary file generator

This first version outputs the tokens of the selected instruments

# Select the instruments to keep by setting the value to True
instruments_to_keep = {
        "distorted0": False,
        "distorted1": False,
        "distorted2": False,
        "distorted3": False,
        "distorted4": False,
        "distorted5": False,
        "clean0": False,
        "clean1": False,
        "clean2": False,
        "clean3": False,
        "clean4": False,
        "bass": True,
        "leads": False,    
        "pads": False,
        "drums": False,
    }

# get_tokens_inst_iter_folders(instruments_to_keep, path_to_general_read_folder, path_to_general_write_folder)

# 8min10s to generate all the token files with only bass and clean0

# 5min30s to generate all the bass tokens without checking if there is bass or not
# 4min49s for bass tokens if we check in the whole file if there is bass or not
# 6min37s to generate the bass tokens when checking at track-level the presence of bass (11 230 tracks without bass)

# 8min12s to generate the token files with the guitars (clean and distorted)

### Automatic computation of the rythmic instrument's tokens (except bass)

The goal is to iterate over the paths in df_rg_3, and select the tokens of the instruments that are track rythmic.
Generally there should be only one track_rythmic instrument per song, but there can be several if they had the same proportion of rythmic measures over the track.

df_rg_3 = pd.read_csv('../../data/df_rg_3.csv')

# Remove measure level
df_rg_3_track_level = df_rg_3.drop_duplicates(subset=['Dadagp_Name', 'Dadagp_Path', 'File_Name'])

print(df_rg_3_track_level.shape) # One row for each part of each track

df_rg_3_track_level.head()

df_rg_3_track_level[df_rg_3_track_level['is_track_rythmic']]

# Only keep the rythmic part of each track?

if windows11:
    # some paths start with a whitespace but windows11 automatically removes those so they have to be changed in the dataframe
    # for instance 
    # '..\\..\\data\\DadaGP\\DadaGP-v1.1\\U\\Unknown\\ - Isolated.gp4.tokens.txt'
    # has to be replaced with
    # '..\\..\\data\\DadaGP\\DadaGP-v1.1\\U\\Unknown\\- Isolated.gp4.tokens.txt'
    aux = df_rg_3_track_level[df_rg_3_track_level['Dadagp_Path'].str.contains(r'\\ -')]['Dadagp_Path']
    aux_list = aux.tolist()
    df_rg_3_track_level.loc[:, 'Dadagp_Path'] = df_rg_3_track_level['Dadagp_Path'].replace(
    aux_list, [x.replace(' -', '-') for x in aux_list])
    
    print(aux_list)
df_rg_3_track_level.describe()

instruments_to_keep

get_tokens_inst_iter_df_rg(df_rg_3_track_level, path_to_general_read_folder, path_to_general_write_folder)

# BGTG dataset is generated!
