==> src/1. preprocessing/6.cp_drums_post_processing.ipynb <==
## Post processing

The goal of this notebook is to pass our data through the post_processing.py script from cp_drums to get a similar form of the data

# Import the autoreload extension
%load_ext autoreload

# Set autoreload to reload all modules (except those excluded by %aimport)
%autoreload 2

import pickle5 as pickle5 #for older python versions you may need pickle5
import pickle
import numpy as np
from random import shuffle
from aux_fun import create_onehot_dict
from tqdm import tqdm

data_path = '..\..\data\processed\preprocessed_dadagp_16_8_800_50.pickle'
parameters = data_path.split('_')[2:6]
parameters = '_'.join(parameters)
parameters = parameters.split('.')[0]
parameters

"""1.Post process to onehot encoding dictionaries"""

with open(data_path, 'rb') as handle:
    fDict = pickle5.load(handle)

#calculate occurences (vocab sizes) for each CP stream and max encoder-decoder 
#sequence lengths

max_enc_length = 0 
max_dec_length = 0 

allEnc_Occs = [] #Encoder

allDec_Occs = [] #Decoder 

for k in range(0, len(fDict['Encoder_RG'])):
    
    #get max seq_lengths
    if max_enc_length < len(fDict['Encoder_RG'][k]):
        max_enc_length = len(fDict['Encoder_RG'][k])
    if max_dec_length < len(fDict['Decoder_Bass'][k]):
        max_dec_length = len(fDict['Decoder_Bass'][k])
    
    #get allEncoder and Decoder events and store them to the lists
    allEnc_Occs.extend(list(set(fDict['Encoder_RG'][k])))
    allDec_Occs.extend(list(set(fDict['Decoder_Bass'][k])))
 
        
#Add in the vocabulories the EOS SOS flags Parallel
allEnc_Occs.extend(['sos','eos'])
allDec_Occs.extend(['sos','eos'])
#Create one-hot dictionaries
Enc_Encoder = create_onehot_dict(allEnc_Occs)
Dec_Encoder = create_onehot_dict(allDec_Occs)

#vocabulory sizes
enc_vocab = Enc_Encoder.categories_[0].shape[0]
dec_vocab = Dec_Encoder.categories_[0].shape[0]


#save the Encoders for the generation stage
encoders_path = r'..\..\data\processed\bass_encoders_cp.pickle'
with open(encoders_path, 'wb') as handle:
    pickle.dump([Enc_Encoder, Dec_Encoder], 
                handle, protocol=pickle.HIGHEST_PROTOCOL)


# '''2. Transform the dictionaries to one-hot encodings and add padding'''

# #set sequence length encoder decoder 
# dec_seq_length = max_dec_length + 1 #for sos or eos #545
# enc_seq_length = max_enc_length + 2 #for sos and eos indications #597



# trainDict = {'All_Events': [],
#               'Encoder_Input': [],
#             'Decoder_Input': [],
#             'Decoder_Output': []}


# for t in tqdm(range(0, len(fDict['Encoder_RG']))):
#     #store All_Events for later use
#     allEvents_seq = fDict['All_Events'][t]
#     trainDict['All_Events'].append(allEvents_seq)
    
#     #prepare data for encoders decoders CP
#     aEnc_seq = fDict['Encoder_RG'][t]
    
#     aDec_seq = fDict['Decoder_Bass'][t]
      
#     pad_lgt_enc_P = enc_seq_length-len(aEnc_seq)-2 #calculate paddings
#     pad_lgt_dec_P = dec_seq_length-len(aDec_seq)-1 #same for both outputs

    
#     '''Encoder'''
#     Enc_pad_emb = np.array(pad_lgt_enc_P*[0])   
    
#     Enc_Input = Enc_Encoder.transform(np.array(['sos']+aEnc_seq+['eos']).reshape(-1, 1)).toarray()
#     Enc_Input = [np.where(r==1)[0][0] for r in Enc_Input] #for embeddings
#     Enc_Input = [x+1 for x in Enc_Input] #shift by one in order to have 0 as pad
#     trainDict['Encoder_Input'].append(np.concatenate((Enc_Input,Enc_pad_emb), axis = 0))
    
#     '''Decoder'''
#     Dec_pad_emb = np.array(pad_lgt_dec_P*[0]) 
    
#     Dec_Input = Dec_Encoder.transform(np.array(['sos']+aDec_seq).reshape(-1, 1)).toarray()
#     Dec_Input = [np.where(r==1)[0][0] for r in Dec_Input] 
#     Dec_Input = [x+1 for x in Dec_Input] 
#     trainDict['Decoder_Input'].append(np.concatenate((Dec_Input,Dec_pad_emb), axis = 0))  

#     Dec_Tf = Dec_Encoder.transform(np.array(aDec_seq+['eos']).reshape(-1, 1)).toarray()
#     Dec_Tf = [np.where(r==1)[0][0] for r in Dec_Tf] 
#     Dec_Tf = [x+1 for x in Dec_Tf] 
#     trainDict['Decoder_Output'].append(np.concatenate((Dec_Tf, Dec_pad_emb), axis = 0)) 
    

'''2. Transform the dictionaries to one-hot encodings and add padding'''

#set sequence length encoder decoder 
dec_seq_length = max_dec_length + 1 #for sos or eos #545
enc_seq_length = max_enc_length + 2 #for sos and eos indications #597



trainDict = {'All_Events': [],
            'Encoder_Input': [],
            'Decoder_Input': [],
            'Decoder_Output': []}

# Precompute padding arrays for efficiency
Enc_pad_emb = np.zeros(enc_seq_length)  # Encoder padding template
Dec_pad_emb = np.zeros(dec_seq_length)  # Decoder padding template

for t in tqdm(range(len(fDict['Encoder_RG']))):
    # Store All_Events for later use
    trainDict['All_Events'].append(fDict['All_Events'][t])

    # Extract sequences
    aEnc_seq = fDict['Encoder_RG'][t]
    aDec_seq = fDict['Decoder_Bass'][t]

    # Compute padding lengths
    pad_lgt_enc_P = enc_seq_length - len(aEnc_seq) - 2
    pad_lgt_dec_P = dec_seq_length - len(aDec_seq) - 1

    ''' Encoder Processing '''
    enc_input_seq = np.array(['sos'] + aEnc_seq + ['eos']).reshape(-1, 1)
    Enc_Input = Enc_Encoder.transform(enc_input_seq).toarray()
    Enc_Input = np.argmax(Enc_Input, axis=1) + 1  # Convert one-hot to indices and shift by 1
    
    # Apply padding
    Enc_Padded = np.concatenate((Enc_Input, Enc_pad_emb[:pad_lgt_enc_P]))
    trainDict['Encoder_Input'].append(Enc_Padded)

    ''' Decoder Processing '''
    dec_input_seq = np.array(['sos'] + aDec_seq).reshape(-1, 1)
    Dec_Input = Dec_Encoder.transform(dec_input_seq).toarray()
    Dec_Input = np.argmax(Dec_Input, axis=1) + 1  

    dec_tf_seq = np.array(aDec_seq + ['eos']).reshape(-1, 1)
    Dec_Tf = Dec_Encoder.transform(dec_tf_seq).toarray()
    Dec_Tf = np.argmax(Dec_Tf, axis=1) + 1  

    # Apply padding
    Dec_Input_Padded = np.concatenate((Dec_Input, Dec_pad_emb[:pad_lgt_dec_P]))
    Dec_Tf_Padded = np.concatenate((Dec_Tf, Dec_pad_emb[:pad_lgt_dec_P]))

    trainDict['Decoder_Input'].append(Dec_Input_Padded)
    trainDict['Decoder_Output'].append(Dec_Tf_Padded)

print(enc_seq_length, dec_seq_length, enc_vocab, dec_vocab)

'''Split the dataset to train test val 80-10-10'''
index_shuf = list(range(len(trainDict['Encoder_Input']))) #random shufling
shuffle(index_shuf)

trainSet = {'All_Events': [],
            'Encoder_Input': [],
            'Decoder_Input': [],
            'Decoder_Output': []}

valSet = {'All_Events': [],
            'Encoder_Input': [],
            'Decoder_Input': [],
            'Decoder_Output': []}


testSet = {'All_Events': [],
            'Encoder_Input': [],
            'Decoder_Input': [],
            'Decoder_Output': []}


trIDXs = int(0.8*len(index_shuf))
trIDX_t_vs = int(0.9*len(index_shuf))

for i in range(0, trIDXs):
    trainSet['All_Events'].append(trainDict['All_Events'][index_shuf[i]])
    trainSet['Encoder_Input'].append(trainDict['Encoder_Input'][index_shuf[i]])
    trainSet['Decoder_Input'].append(trainDict['Decoder_Input'][index_shuf[i]])
    trainSet['Decoder_Output'].append(trainDict['Decoder_Output'][index_shuf[i]])

for i in range(trIDXs, trIDX_t_vs):
    testSet['All_Events'].append(trainDict['All_Events'][index_shuf[i]])
    testSet['Encoder_Input'].append(trainDict['Encoder_Input'][index_shuf[i]])
    testSet['Decoder_Input'].append(trainDict['Decoder_Input'][index_shuf[i]])
    testSet['Decoder_Output'].append(trainDict['Decoder_Output'][index_shuf[i]])
    
for i in range(trIDX_t_vs, len(index_shuf)):
    valSet['All_Events'].append(trainDict['All_Events'][index_shuf[i]])
    valSet['Encoder_Input'].append(trainDict['Encoder_Input'][index_shuf[i]])
    valSet['Decoder_Input'].append(trainDict['Decoder_Input'][index_shuf[i]])
    valSet['Decoder_Output'].append(trainDict['Decoder_Output'][index_shuf[i]])


#save them
train_path = r"..\..\data\processed\train_set_streams_" + parameters + ".pickle"
test_path = r"..\..\data\processed\test_set_streams_" + parameters + ".pickle"
val_path = r"..\..\data\processed\val_set_streams_" + parameters + ".pickle"

with open(train_path, 'wb') as handle:
    pickle.dump(trainSet, handle, protocol=pickle.HIGHEST_PROTOCOL)
    
with open(test_path, 'wb') as handle:
    pickle.dump(testSet, handle, protocol=pickle.HIGHEST_PROTOCOL)

with open(val_path, 'wb') as handle:
    pickle.dump(valSet, handle, protocol=pickle.HIGHEST_PROTOCOL)
